refactor(api): extrai validators enterprise-cloud + probe compartilhado de validation.ts
Decomposição god-file (Tier 1 / #3501), 3ª fatia. Move os 10 validators de
provider enterprise-cloud + o probe de chat compartilhado para módulos
co-localizados, byte-idênticos ao inline original.

- providers/validation/directChatProbe.ts — validateDirectChatProvider (probe POST
  /chat/completions compartilhado por command-code, nlpcloud e os cloud validators;
  extraído p/ leaf próprio p/ host e cloudProviders dividirem sem ciclo)
- providers/validation/cloudProviders.ts  — heroku/databricks/datarobot/snowflake/
  gigachat/azure-openai/azure-ai/watsonx/oci/sap

validation.ts: 3027 → 2333 (−694; acumulado −2189 das 3 fatias). Imports órfãos
removidos do host: gigachatAuth, azureAi/datarobot/oci/sap/watsonx (config), e 5
normalize* do urlHelpers.

Validação: typecheck:core, eslint (0/0), check:cycles, check:file-size (módulos
<800), 154 testes de validação existentes + 3 de caracterização — todos verdes.
